+++ src/index.ts
@@ -46,8 +46,10 @@ export { default as SwipeActions } from './components/data/SwipeActions.vue'
 // Navigation
 export { default as NavigationStack } from './components/navigation/NavigationStack.vue'
 export { default as NavigationLink } from './components/navigation/NavigationLink.vue'
+export { default as NavigationSplitView } from './components/navigation/NavigationSplitView.vue'
 export { default as TabView } from './components/navigation/TabView.vue'
 export { default as Sheet } from './components/navigation/Sheet.vue'
+export { default as FullScreenCover } from './components/navigation/FullScreenCover.vue'
 
 // Feedback
 export { default as Alert } from './components/feedback/SVAlert.vue'
@@ -66,6 +68,8 @@ export type { MenuAction, MenuProps } from './components/controls/Menu.vue'
 export type { ContextMenuProps } from './components/controls/ContextMenu.vue'
 export type { GaugeProps } from './components/feedback/Gauge.vue'
 export type { NavigationStackProps } from './components/navigation/NavigationStack.vue'
+export type { NavigationSplitViewProps, SplitViewVisibility } from './components/navigation/NavigationSplitView.vue'
+export type { FullScreenCoverProps } from './components/navigation/FullScreenCover.vue'
 export type { ListProps } from './components/data/SVList.vue'
 export type { TransitionPreset } from './components/motion/TransitionView.vue'